In CLI support setting TLS policy to any known type
Previously you could only do either a file or the default policy,
and tls_proxy was hardcoded to only do the default policy.
